requirements.txt
python-telegram-bot==10.0.1
requests==2.18.1
urllib3==1.21.1
beautifulsoup4
python-dotenv
apiai==1.2.3
geopy
emoji
pyflight
selenium